+++ src/config-provider/abstract/config-provider-abstract.module.ts
@@ -5,7 +5,7 @@ import {
   Provider,
 } from '@nestjs/common';
 import { reloadableSourceToken } from './config-provider-tokens';
-import { CONFIG_PROVIDER_ERRORS } from './config-provider-error-codes';
+import { CONFIG_PROVIDER_ERRORS } from './config-provider.error';
 import { ConfigProviderError } from './config-provider.error';
 import {
   ConfigProviderModuleAsyncOptions,
