+++ src/config-provider/env-adapter/env-config.adapter.unit.spec.ts
@@ -1,7 +1,7 @@
 import * as fs from 'node:fs';
 import * as dotenv from 'dotenv';
 import { EnvConfigAdapter } from './env-config.adapter';
-import { CONFIG_PROVIDER_ERRORS } from '../abstract/config-provider-error-codes';
+import { CONFIG_PROVIDER_ERRORS } from '../abstract/config-provider.error';
 import { ConfigProviderError } from '../abstract/config-provider.error';
 
 jest.mock('node:fs');
