+++ src/config-provider/secrets-manager-adapter/secrets-manager-config.adapter.unit.spec.ts
@@ -1,6 +1,6 @@
 import { SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
 import { SecretsManagerConfigAdapter } from './secrets-manager-config.adapter';
-import { CONFIG_PROVIDER_ERRORS } from '../abstract/config-provider-error-codes';
+import { CONFIG_PROVIDER_ERRORS } from '../abstract/config-provider.error';
 import { ConfigProviderError } from '../abstract/config-provider.error';
 
 const mockSend = jest.fn();
